script: Move CSS DOM interfaces to `script/dom/css/`
