+++ components/script/devtools.rs
@@ -35,8 +35,8 @@ use crate::dom::bindings::conversions::{ConversionResult, FromJSValConvertible};
 use crate::dom::bindings::inheritance::Castable;
 use crate::dom::bindings::root::DomRoot;
 use crate::dom::bindings::str::DOMString;
-use crate::dom::cssstyledeclaration::ENABLED_LONGHAND_PROPERTIES;
-use crate::dom::cssstylerule::CSSStyleRule;
+use crate::dom::css::cssstyledeclaration::ENABLED_LONGHAND_PROPERTIES;
+use crate::dom::css::cssstylerule::CSSStyleRule;
 use crate::dom::document::AnimationFrameCallback;
 use crate::dom::element::Element;
 use crate::dom::globalscope::GlobalScope;
