+++ components/script/dom/css/csskeyframesrule.rs
@@ -12,16 +12,16 @@ use style::stylesheets::keyframes_rule::{Keyframe, KeyframeSelector, KeyframesRu
 use style::stylesheets::{CssRuleType, StylesheetInDocument};
 use style::values::KeyframesName;
 
+use super::csskeyframerule::CSSKeyframeRule;
+use super::cssrule::{CSSRule, SpecificCSSRule};
+use super::cssrulelist::{CSSRuleList, RulesSource};
+use super::cssstylesheet::CSSStyleSheet;
 use crate::dom::bindings::codegen::Bindings::CSSKeyframesRuleBinding::CSSKeyframesRuleMethods;
 use crate::dom::bindings::error::ErrorResult;
 use crate::dom::bindings::inheritance::Castable;
 use crate::dom::bindings::reflector::{DomGlobal, reflect_dom_object};
 use crate::dom::bindings::root::{DomRoot, MutNullableDom};
 use crate::dom::bindings::str::DOMString;
-use crate::dom::csskeyframerule::CSSKeyframeRule;
-use crate::dom::cssrule::{CSSRule, SpecificCSSRule};
-use crate::dom::cssrulelist::{CSSRuleList, RulesSource};
-use crate::dom::cssstylesheet::CSSStyleSheet;
 use crate::dom::window::Window;
 use crate::script_runtime::CanGc;
