+++ components/script/dom/css/cssrulelist.rs
@@ -17,15 +17,15 @@ use style::stylesheets::{
     StylesheetInDocument, StylesheetLoader as StyleStylesheetLoader,
 };
 
+use super::csskeyframerule::CSSKeyframeRule;
+use super::cssrule::CSSRule;
+use super::cssstylesheet::CSSStyleSheet;
 use crate::conversions::Convert;
 use crate::dom::bindings::cell::DomRefCell;
 use crate::dom::bindings::codegen::Bindings::CSSRuleListBinding::CSSRuleListMethods;
 use crate::dom::bindings::error::{Error, ErrorResult, Fallible};
 use crate::dom::bindings::reflector::{DomGlobal, Reflector, reflect_dom_object};
 use crate::dom::bindings::root::{Dom, DomRoot, MutNullableDom};
-use crate::dom::csskeyframerule::CSSKeyframeRule;
-use crate::dom::cssrule::CSSRule;
-use crate::dom::cssstylesheet::CSSStyleSheet;
 use crate::dom::html::htmlelement::HTMLElement;
 use crate::dom::window::Window;
 use crate::script_runtime::CanGc;
