+++ components/script/dom/css/cssstylesheet.rs
@@ -18,6 +18,9 @@ use style::stylesheets::{
     StylesheetInDocument, UrlExtraData,
 };
 
+use super::cssrulelist::{CSSRuleList, RulesSource};
+use super::stylesheet::StyleSheet;
+use super::stylesheetlist::StyleSheetListOwner;
 use crate::dom::bindings::cell::DomRefCell;
 use crate::dom::bindings::codegen::Bindings::CSSStyleSheetBinding::{
     CSSStyleSheetInit, CSSStyleSheetMethods,
@@ -32,14 +35,11 @@ use crate::dom::bindings::reflector::{
 };
 use crate::dom::bindings::root::{DomRoot, MutNullableDom};
 use crate::dom::bindings::str::{DOMString, USVString};
-use crate::dom::cssrulelist::{CSSRuleList, RulesSource};
 use crate::dom::document::Document;
 use crate::dom::element::Element;
 use crate::dom::html::htmlstyleelement::HTMLStyleElement;
 use crate::dom::medialist::MediaList;
 use crate::dom::node::NodeTraits;
-use crate::dom::stylesheet::StyleSheet;
-use crate::dom::stylesheetlist::StyleSheetListOwner;
 use crate::dom::types::Promise;
 use crate::dom::window::Window;
 use crate::script_runtime::CanGc;
