+++ components/script/dom/css/mod.rs
@@ -1,0 +1,32 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#[allow(clippy::module_inception, reason = "The interface name is CSS")]
+pub(crate) mod css;
+pub(crate) mod cssconditionrule;
+pub(crate) mod cssfontfacerule;
+pub(crate) mod cssgroupingrule;
+pub(crate) mod cssimportrule;
+pub(crate) mod csskeyframerule;
+pub(crate) mod csskeyframesrule;
+pub(crate) mod csslayerblockrule;
+pub(crate) mod csslayerstatementrule;
+pub(crate) mod cssmediarule;
+pub(crate) mod cssnamespacerule;
+pub(crate) mod cssnesteddeclarations;
+pub(crate) mod cssrule;
+pub(crate) mod cssrulelist;
+pub(crate) mod cssstyledeclaration;
+pub(crate) mod cssstylerule;
+pub(crate) mod cssstylesheet;
+pub(crate) mod cssstylevalue;
+pub(crate) mod csssupportsrule;
+pub(crate) mod fontface;
+pub(crate) mod fontfaceset;
+pub(crate) mod stylepropertymapreadonly;
+pub(crate) mod stylesheet;
+pub(crate) mod stylesheetcontentscache;
+pub(crate) mod stylesheetlist;
+
+pub(crate) use self::css::CSS;
