+++ components/script/dom/mod.rs
@@ -251,24 +251,7 @@ pub(crate) mod cryptokey;
 pub(crate) mod csp;
 pub(crate) mod csppolicyviolationreport;
 pub(crate) mod css;
-pub(crate) mod cssconditionrule;
-pub(crate) mod cssfontfacerule;
-pub(crate) mod cssgroupingrule;
-pub(crate) mod cssimportrule;
-pub(crate) mod csskeyframerule;
-pub(crate) mod csskeyframesrule;
-pub(crate) mod csslayerblockrule;
-pub(crate) mod csslayerstatementrule;
-pub(crate) mod cssmediarule;
-pub(crate) mod cssnamespacerule;
-pub(crate) mod cssnesteddeclarations;
-pub(crate) mod cssrule;
-pub(crate) mod cssrulelist;
-pub(crate) mod cssstyledeclaration;
-pub(crate) mod cssstylerule;
-pub(crate) mod cssstylesheet;
-pub(crate) mod cssstylevalue;
-pub(crate) mod csssupportsrule;
+pub(crate) use self::css::*;
 pub(crate) mod customelementregistry;
 pub(crate) mod customevent;
 pub(crate) mod customstateset;
@@ -320,8 +303,6 @@ pub(crate) mod filelist;
 pub(crate) mod filereader;
 pub(crate) mod filereadersync;
 pub(crate) mod focusevent;
-pub(crate) mod fontface;
-pub(crate) mod fontfaceset;
 pub(crate) mod formdata;
 pub(crate) mod formdataevent;
 pub(crate) mod gamepad;
@@ -432,10 +413,6 @@ pub(crate) mod shadowroot;
 pub(crate) mod staticrange;
 pub(crate) mod storage;
 pub(crate) mod storageevent;
-pub(crate) mod stylepropertymapreadonly;
-pub(crate) mod stylesheet;
-pub(crate) mod stylesheetcontentscache;
-pub(crate) mod stylesheetlist;
 pub(crate) mod submitevent;
 pub(crate) mod subtlecrypto;
 pub(crate) mod svg;
